+++ crates/mesh-mixture-of-agents/src/lib.rs
@@ -4,31 +4,52 @@
 //! outputs with deterministic logic, and return one coherent OpenAI-
 //! compatible response.  The client thinks it talks to one model.
 //!
-//! Transport is abstracted behind the [`ModelBackend`] trait:
-//! - Default [`HttpBackend`] talks to any OpenAI-compatible HTTP endpoint
-//! - The mesh host-runtime provides a mesh-native backend that dispatches
-//!   local models via direct HTTP and remote models via QUIC tunnel
+//! Transport is abstracted behind the [`ModelBackend`] trait (see
+//! [`backend`]). The default [`HttpBackend`] talks to any
+//! OpenAI-compatible HTTP endpoint and is suitable for standalone/test
+//! use. The mesh host-runtime provides mesh-native backends that
+//! dispatch local models via direct HTTP and remote models via QUIC
+//! tunnel.
 //!
 //! ```text
 //! Agent / Goose / pi
 //!     │
 //!     │  POST /v1/chat/completions { "model": "mesh" }
 //!     ▼
-//!  MoA Gateway
-//!   ├─ session / context packing (role-shaped)
-//!   ├─ parallel fan-out via ModelBackend
-//!   ├─ incremental gathering with early-exit on consensus
-//!   ├─ deterministic arbiter (code, not models)
-//!   └─ reducer escalation only on genuine conflict
+//!  MoA Gateway  (handle_turn)
+//!   ├─ session / context packing (role-shaped)        — context::*
+//!   ├─ parallel fan-out via ModelBackend              — fanout::gather_workers_incremental
+//!   ├─ incremental gathering with early-exit          — arbiter::try_early_decision
+//!   ├─ deterministic arbiter (code, not models)       — arbiter::arbitrate
+//!   └─ reducer escalation only on genuine conflict    — reducer::hedged_reducer_call
 //! ```
+//!
+//! Modules:
+//! - [`backend`] — `ModelBackend` trait, `HttpBackend`, `SamplingParams`,
+//!   `ModelEntry`
+//! - [`reducer`] — reducer candidate ordering, hedged ladder
+//! - [`fanout`] — incremental worker gathering with early-exit
+//! - [`arbiter`] — deterministic arbitration + early-exit consensus
+//! - [`normalize`] — 3-tier dirty-output parsing
+//! - [`session`] — canonical transcript + turn classification
+//! - [`context`] — role-shaped context packing
+//! - [`worker`] — role assignment, think-tag stripping
 
 pub mod arbiter;
+pub mod backend;
 pub mod context;
+mod fanout;
 pub mod normalize;
+mod reducer;
 pub mod session;
 pub mod worker;
 
+pub use backend::{HttpBackend, ModelBackend, ModelEntry, SamplingParams};
+
+use backend::call_backend;
+use fanout::gather_workers_incremental;
 use normalize::WorkerOutput;
+use reducer::{hedged_reducer_call, reducer_candidates};
 use serde_json::{json, Value};
 use session::Session;
 use std::time::{Duration, Instant};
@@ -36,135 +57,6 @@ use worker::WorkerRole;
 
 /// The virtual model name that triggers MoA routing.
 pub const VIRTUAL_MODEL_NAME: &str = "mesh";
-
-// ─── Sampling params ─────────────────────────────────────────────────
-
-/// Sampling hyperparameters sent to backend models.
-/// Workers get higher temperature for diversity; reducer gets lower for precision.
-#[derive(Debug, Clone, Copy)]
-pub struct SamplingParams {
-    pub temperature: f32,
-    pub top_p: f32,
-}
-
-impl SamplingParams {
-    /// High-diversity settings for MoA workers — encourages each model
-    /// to explore different parts of the solution space.
-    pub fn worker() -> Self {
-        Self {
-            temperature: 0.8,
-            top_p: 0.95,
-        }
-    }
-
-    /// Low-variance settings for the reducer — precise synthesis.
-    pub fn reducer() -> Self {
-        Self {
-            temperature: 0.3,
-            top_p: 0.9,
-        }
-    }
-}
-
-impl Default for SamplingParams {
-    fn default() -> Self {
-        Self::reducer()
-    }
-}
-
-// ─── Backend trait ───────────────────────────────────────────────────
-
-/// Abstraction for calling a model.  The gateway doesn't care whether
-/// the model is local HTTP, remote QUIC, or something else entirely.
-#[async_trait::async_trait]
-pub trait ModelBackend: Send + Sync + 'static {
-    /// Call the model with the given messages (and optionally tools).
-    /// Returns the full JSON response body from the model.
-    async fn chat_completion(
-        &self,
-        model: &str,
-        messages: &[Value],
-        tools: Option<&Value>,
-        max_tokens: u32,
-        timeout: Duration,
-        sampling: SamplingParams,
-    ) -> Result<Value, String>;
-}
-
-/// Default HTTP backend — calls any OpenAI-compatible endpoint.
-pub struct HttpBackend {
-    pub base_url: String,
-    http: reqwest::Client,
-}
-
-impl HttpBackend {
-    pub fn new(base_url: String) -> Self {
-        let http = reqwest::Client::builder()
-            .timeout(Duration::from_secs(120))
-            .build()
-            .unwrap_or_default();
-        Self { base_url, http }
-    }
-}
-
-#[async_trait::async_trait]
-impl ModelBackend for HttpBackend {
-    async fn chat_completion(
-        &self,
-        model: &str,
-        messages: &[Value],
-        tools: Option<&Value>,
-        max_tokens: u32,
-        timeout: Duration,
-        sampling: SamplingParams,
-    ) -> Result<Value, String> {
-        let url = format!("{}/chat/completions", self.base_url);
-        let mut body = json!({
-            "model": model,
-            "messages": messages,
-            "max_tokens": max_tokens,
-            "temperature": sampling.temperature,
-            "top_p": sampling.top_p,
-            "stream": false,
-        });
-        if let Some(tools) = tools {
-            body.as_object_mut()
-                .unwrap()
-                .insert("tools".to_string(), tools.clone());
-        }
-
-        let resp = self
-            .http
-            .post(&url)
-            .json(&body)
-            .timeout(timeout)
-            .send()
-            .await
-            .map_err(|e| format!("request failed: {e}"))?;
-
-        let status = resp.status();
-        if !status.is_success() {
-            let text = resp.text().await.unwrap_or_default();
-            return Err(format!("HTTP {status}: {}", &text[..text.len().min(200)]));
-        }
-
-        resp.json::<Value>()
-            .await
-            .map_err(|e| format!("response parse: {e}"))
-    }
-}
-
-// ─── Model entry ─────────────────────────────────────────────────────
-
-/// A model available for MoA fan-out.
-#[derive(Clone)]
-pub struct ModelEntry {
-    /// Model name (as used in the API).
-    pub name: String,
-    /// Index into the backends vec.  Multiple models can share a backend
-    /// (e.g. all models behind the same proxy) or each have their own.
-    pub backend_index: usize,
-}
 
 // ─── Configuration ───────────────────────────────────────────────────
 
@@ -263,7 +155,7 @@ pub async fn handle_turn(config: &GatewayConfig, body: &Value) -> TurnResult {
 /// If `allowed_tools` is empty (no tools were declared on the request
 /// body) we leave the proposal as-is — the arbiter will route to the
 /// reducer which has the same call site policy.
-fn enforce_allowed_tools(
+pub(crate) fn enforce_allowed_tools(
     output: &mut normalize::WorkerOutput,
     allowed_tools: &[String],
     model: &str,
@@ -519,379 +411,6 @@ async fn resolve_decision(
     }
 }
 
-/// Pick the reducer — prefers first model (typically local, zero RTT).
-/// Reducer candidates in priority order: big-tier models first (multi-
-/// digit B, or names with no size like MiniMax), then small-tier models
-/// as last-resort fallback. Callers should try each in order and stop
-/// on the first that succeeds, so a broken big-tier peer (e.g. a peer
-/// running a stale binary that 502s on tool calls) doesn't take down
-/// the whole reducer step.
-fn reducer_candidates(config: &GatewayConfig) -> Vec<(String, usize)> {
-    let mut big = Vec::new();
-    let mut small = Vec::new();
-    for m in &config.models {
-        let entry = (m.name.clone(), m.backend_index);
-        if worker::is_single_digit_b_name(&m.name) {
-            small.push(entry);
-        } else {
-            big.push(entry);
-        }
-    }
-    big.extend(small);
-    if big.is_empty() {
-        big.push(("unknown".into(), 0));
-    }
-    big
-}
-
-/// Call the ordered reducer candidates with hedging.
-///
-/// Starts the first candidate immediately. If it hasn't returned within
-/// `hedge_delay`, the next candidate is started in parallel without
-/// cancelling the in-flight one — we race for the first OK. If a candidate
-/// errors, the next one is started immediately (no hedge wait).
-///
-/// Returns the first successful (name, response_text). If every candidate
-/// fails, returns the last error encountered.
-///
-/// Cost shape:
-/// - Happy path (cand 0 OK in <hedge_delay): exactly 1 backend call.
-/// - Slow happy path (cand 0 OK in hedge_delay..reducer_timeout): up to 2
-///   overlapping calls, accept whichever wins, cancel the loser.
-/// - Fast-fail (cand 0 errors quickly): immediate move to cand 1, 1 call.
-/// - All fail: at most N calls, capped at reducer_timeout + (N-1)·hedge_delay
-///   end-to-end (vs N·reducer_timeout sequentially).
-async fn hedged_reducer_call(
-    backends: &[std::sync::Arc<dyn ModelBackend>],
-    candidates: Vec<(String, usize)>,
-    messages: Vec<Value>,
-    tools: Option<Value>,
-    timeout: Duration,
-    hedge_delay: Duration,
-) -> Result<(String, String), String> {
-    use tokio::task::JoinSet;
-
-    if candidates.is_empty() {
-        return Err("no reducer candidates".into());
-    }
-
-    let mut join_set: JoinSet<(String, Result<String, String>)> = JoinSet::new();
-    let mut remaining = candidates.into_iter();
-    let mut last_err: Option<String> = None;
-
-    // Spawn a single candidate.
-    fn spawn(
-        join_set: &mut JoinSet<(String, Result<String, String>)>,
-        backends: &[std::sync::Arc<dyn ModelBackend>],
-        name: String,
-        backend_idx: usize,
-        messages: Vec<Value>,
-        tools: Option<Value>,
-        timeout: Duration,
-    ) {
-        let backend = backends[backend_idx].clone();
-        tracing::info!("moa: reducer hedge → {name}");
-        join_set.spawn(async move {
-            let result = call_backend(
-                &*backend,
-                &name,
-                &messages,
-                tools.as_ref(),
-                2048,
-                timeout,
-                SamplingParams::reducer(),
-            )
-            .await;
-            (name, result)
-        });
-    }
-
-    // Start candidate 0.
-    if let Some((name, idx)) = remaining.next() {
-        spawn(
-            &mut join_set,
-            backends,
-            name,
-            idx,
-            messages.clone(),
-            tools.clone(),
-            timeout,
-        );
-    }
-
-    // Race in-flight calls against a hedge timer.
-    while !join_set.is_empty() {
-        let hedge_sleep = tokio::time::sleep(hedge_delay);
-        tokio::pin!(hedge_sleep);
-
-        tokio::select! {
-            // A candidate finished.
-            joined = join_set.join_next() => {
-                match joined {
-                    Some(Ok((name, Ok(text)))) => {
-                        // First success wins. Cancel the rest.
-                        join_set.abort_all();
-                        // Drain so cancellations complete cleanly.
-                        while join_set.join_next().await.is_some() {}
-                        return Ok((name, text));
-                    }
-                    Some(Ok((name, Err(e)))) => {
-                        tracing::warn!(
-                            "moa: reducer {name} failed: {e}, trying next candidate"
-                        );
-                        last_err = Some(e);
-                        // Start the next candidate immediately on failure.
-                        if let Some((next_name, next_idx)) = remaining.next() {
-                            spawn(
-                                &mut join_set,
-                                backends,
-                                next_name,
-                                next_idx,
-                                messages.clone(),
-                                tools.clone(),
-                                timeout,
-                            );
-                        }
-                    }
-                    Some(Err(join_err)) => {
-                        tracing::warn!("moa: reducer task join error: {join_err}");
-                        if let Some((next_name, next_idx)) = remaining.next() {
-                            spawn(
-                                &mut join_set,
-                                backends,
-                                next_name,
-                                next_idx,
-                                messages.clone(),
-                                tools.clone(),
-                                timeout,
-                            );
-                        }
-                    }
-                    None => break,
-                }
-            }
-            // Hedge timer fires: start another candidate alongside in-flight ones.
-            _ = &mut hedge_sleep => {
-                if let Some((next_name, next_idx)) = remaining.next() {
-                    spawn(
-                        &mut join_set,
-                        backends,
-                        next_name,
-                        next_idx,
-                        messages.clone(),
-                        tools.clone(),
-                        timeout,
-                    );
-                }
-                // If no more to start, just wait on the JoinSet without the
-                // hedge timer racing again (next loop iteration's sleep will
-                // simply never fire because we'll take the join branch).
-            }
-        }
-    }
-
-    Err(last_err.unwrap_or_else(|| "all reducer candidates failed".into()))
-}
-
-// ─── Backend call + text extraction ──────────────────────────────────
-
-/// Call a backend and extract the assistant text from the response.
-/// Retries once on HTTP 429 (rate limit) after the server's `retry-after`
-/// delay (default 1s).
-async fn call_backend(
-    backend: &dyn ModelBackend,
-    model: &str,
-    messages: &[Value],
-    tools: Option<&Value>,
-    max_tokens: u32,
-    timeout: Duration,
-    sampling: SamplingParams,
-) -> Result<String, String> {
-    match backend
-        .chat_completion(model, messages, tools, max_tokens, timeout, sampling)
-        .await
-    {
-        Ok(resp) => extract_text_from_response(&resp),
-        Err(e) if e.contains("429") => {
-            // Parse retry-after from error message if present, default 1s
-            let delay = parse_retry_after(&e).unwrap_or(1);
-            tracing::info!("moa: 429 from {model}, retrying after {delay}s");
-            tokio::time::sleep(Duration::from_secs(delay)).await;
-            let resp = backend
-                .chat_completion(model, messages, tools, max_tokens, timeout, sampling)
-                .await?;
-            extract_text_from_response(&resp)
-        }
-        Err(e) => Err(e),
-    }
-}
-
-/// Extract retry-after seconds from an error message containing "retry-after: N".
-fn parse_retry_after(err: &str) -> Option<u64> {
-    let lower = err.to_lowercase();
-    lower
-        .find("retry-after:")
-        .map(|i| &err[i + 12..])
-        .and_then(|s| s.split_whitespace().next())
-        .and_then(|s| s.trim().parse::<u64>().ok())
-}
-
-/// Extract assistant text from a chat completion response body.
-fn extract_text_from_response(resp: &Value) -> Result<String, String> {
-    let message = &resp["choices"][0]["message"];
-
-    // Native tool_calls → KV format for normalizer
-    if let Some(tool_calls) = message.get("tool_calls").and_then(|tc| tc.as_array()) {
-        if let Some(tc) = tool_calls.first() {
-            let name = tc
-                .pointer("/function/name")
-                .and_then(|n| n.as_str())
-                .unwrap_or("unknown");
-            let args = tc
-                .pointer("/function/arguments")
-                .and_then(|a| a.as_str())
-                .unwrap_or("{}");
-            return Ok(format!(
-                "kind: tool_proposal\ntool: {name}\narguments: {args}\nconfidence: 0.9\npayload: calling {name}",
-            ));
-        }
-    }
-
-    let content = message
-        .get("content")
-        .and_then(|c| c.as_str())
-        .unwrap_or("")
-        .to_string();
-
-    let stripped = worker::strip_thinking(&content);
-    if !stripped.is_empty() {
-        return Ok(stripped);
-    }
-
-    let thinking = worker::extract_thinking(&content);
-    if !thinking.is_empty() {
-        return Ok(thinking);
-    }
-
-    let reasoning = message
-        .get("reasoning")
-        .and_then(|r| r.as_str())
-        .unwrap_or("");
-    if !reasoning.is_empty() {
-        return Ok(reasoning.to_string());
-    }
-
-    Err("empty response".into())
-}
-
-// ─── Worker gathering ────────────────────────────────────────────────
-
-async fn gather_workers_incremental(
-    join_set: &mut tokio::task::JoinSet<(String, WorkerRole, Result<String, String>, u64)>,
-    total_workers: usize,
-    has_tools: bool,
-    allowed_tools: &[String],
-) -> (
-    Vec<WorkerOutput>,
-    Vec<WorkerSummary>,
-    Option<arbiter::Decision>,
-) {
-    let mut outputs = Vec::new();
-    let mut summaries = Vec::new();
-    let mut total_finished: usize = 0;
-
-    while let Some(join_result) = join_set.join_next().await {
-        match join_result {
-            Ok((model, role, Ok(text), elapsed)) => {
-                total_finished += 1;
-                let mut normalized =
-                    normalize::normalize_worker_output(&text, &model, role, elapsed);
-                enforce_allowed_tools(&mut normalized, allowed_tools, &model);
-                tracing::info!(
-                    "moa: worker {} ({}) → {:?} conf={:.2} ({}ms, {} chars)",
-                    model,
-                    role.label(),
-                    normalized.kind,
-                    normalized.confidence,
-                    elapsed,
-                    text.len(),
-                );
-                summaries.push(WorkerSummary {
-                    model: model.clone(),
-                    role,
-                    succeeded: true,
-                    elapsed_ms: elapsed,
-                    output_kind: Some(normalized.kind),
-                    confidence: Some(normalized.confidence),
-                });
-                outputs.push(normalized);
-
-                if let Some(decision) =
-                    arbiter::try_early_decision(&outputs, total_workers, total_finished, has_tools)
-                {
-                    join_set.abort_all();
-                    while let Some(leftover) = join_set.join_next().await {
-                        if let Ok((m, r, result, el)) = leftover {
-                            summaries.push(WorkerSummary {
-                                model: m,
-                                role: r,
-                                succeeded: result.is_ok(),
-                                elapsed_ms: el,
-                                output_kind: None,
-                                confidence: None,
-                            });
-                        }
-                    }
-                    return (outputs, summaries, Some(decision));
-                }
-            }
-            Ok((model, role, Err(e), elapsed)) => {
-                total_finished += 1;
-                tracing::warn!(
-                    "moa: worker {} ({}) failed after {}ms: {}",
-                    model,
-                    role.label(),
-                    elapsed,
-                    e,
-                );
-                summaries.push(WorkerSummary {
-                    model,
-                    role,
-                    succeeded: false,
-                    elapsed_ms: elapsed,
-                    output_kind: None,
-                    confidence: None,
-                });
-
-                if let Some(decision) =
-                    arbiter::try_early_decision(&outputs, total_workers, total_finished, has_tools)
-                {
-                    join_set.abort_all();
-                    while let Some(leftover) = join_set.join_next().await {
-                        if let Ok((m, r, result, el)) = leftover {
-                            summaries.push(WorkerSummary {
-                                model: m,
-                                role: r,
-                                succeeded: result.is_ok(),
-                                elapsed_ms: el,
-                                output_kind: None,
-                                confidence: None,
-                            });
-                        }
-                    }
-                    return (outputs, summaries, Some(decision));
-                }
-            }
-            Err(e) => {
-                total_finished += 1;
-                tracing::warn!("moa: worker task panicked: {e}");
-            }
-        }
-    }
-
-    (outputs, summaries, None)
-}
-
 // ─── Endpoint discovery (for standalone/test use) ────────────────────
 
 /// An endpoint for the HTTP backend (convenience for test harnesses).
@@ -1023,245 +542,4 @@ fn short_id() -> String {
         .unwrap_or_default()
         .as_nanos();
     format!("{:x}", t)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn parse_retry_after_from_header() {
-        let err = "HTTP 429: retry-after: 2\r\ncontent-type: application/json";
-        assert_eq!(parse_retry_after(err), Some(2));
-    }
-
-    #[test]
-    fn parse_retry_after_missing() {
-        let err = "HTTP 429: Too Many Requests";
-        assert_eq!(parse_retry_after(err), None);
-    }
-
-    #[test]
-    fn parse_retry_after_case_insensitive() {
-        let err = "Retry-After: 5";
-        assert_eq!(parse_retry_after(err), Some(5));
-    }
-
-    #[test]
-    fn worker_sampling_high_diversity() {
-        let s = SamplingParams::worker();
-        assert!(s.temperature > 0.5, "workers need high temp for diversity");
-        assert!(s.top_p > 0.9, "workers need high top_p for diversity");
-    }
-
-    #[test]
-    fn reducer_sampling_low_variance() {
-        let s = SamplingParams::reducer();
-        assert!(s.temperature <= 0.4, "reducer needs low temp for precision");
-        assert!(s.top_p <= 0.95, "reducer needs bounded top_p");
-    }
-
-    #[test]
-    fn default_sampling_is_reducer() {
-        let d = SamplingParams::default();
-        let r = SamplingParams::reducer();
-        assert!((d.temperature - r.temperature).abs() < f32::EPSILON);
-        assert!((d.top_p - r.top_p).abs() < f32::EPSILON);
-    }
-
-    // ── hedged_reducer_call ──────────────────────────────────────────
-
-    use std::sync::atomic::{AtomicUsize, Ordering};
-
-    #[derive(Clone)]
-    enum FakeBehavior {
-        OkAfter(Duration, String),
-        ErrAfter(Duration, String),
-    }
-
-    struct FakeBackend {
-        behaviors: std::sync::Mutex<std::collections::HashMap<String, FakeBehavior>>,
-        calls: AtomicUsize,
-    }
-
-    impl FakeBackend {
-        fn new(behaviors: Vec<(&str, FakeBehavior)>) -> std::sync::Arc<Self> {
-            let mut map = std::collections::HashMap::new();
-            for (n, b) in behaviors {
-                map.insert(n.to_string(), b);
-            }
-            std::sync::Arc::new(FakeBackend {
-                behaviors: std::sync::Mutex::new(map),
-                calls: AtomicUsize::new(0),
-            })
-        }
-        fn calls(&self) -> usize {
-            self.calls.load(Ordering::SeqCst)
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl ModelBackend for FakeBackend {
-        async fn chat_completion(
-            &self,
-            model: &str,
-            _messages: &[Value],
-            _tools: Option<&Value>,
-            _max_tokens: u32,
-            _timeout: Duration,
-            _sampling: SamplingParams,
-        ) -> Result<Value, String> {
-            self.calls.fetch_add(1, Ordering::SeqCst);
-            let behavior = self.behaviors.lock().unwrap().get(model).cloned();
-            match behavior {
-                Some(FakeBehavior::OkAfter(d, body)) => {
-                    tokio::time::sleep(d).await;
-                    Ok(serde_json::json!({
-                        "choices": [{"message": {"content": body}}],
-                    }))
-                }
-                Some(FakeBehavior::ErrAfter(d, msg)) => {
-                    tokio::time::sleep(d).await;
-                    Err(msg)
-                }
-                None => Err(format!("unconfigured model: {model}")),
-            }
-        }
-    }
-
-    #[tokio::test]
-    async fn hedged_reducer_happy_path_calls_only_first() {
-        let fake = FakeBackend::new(vec![
-            (
-                "alpha",
-                FakeBehavior::OkAfter(Duration::from_millis(50), "alpha-resp".into()),
-            ),
-            (
-                "beta",
-                FakeBehavior::OkAfter(Duration::from_millis(50), "beta-resp".into()),
-            ),
-        ]);
-        let backends: Vec<std::sync::Arc<dyn ModelBackend>> = vec![fake.clone(), fake.clone()];
-        let candidates = vec![("alpha".into(), 0), ("beta".into(), 1)];
-
-        let res = hedged_reducer_call(
-            &backends,
-            candidates,
-            vec![],
-            None,
-            Duration::from_secs(15),
-            Duration::from_secs(5),
-        )
-        .await;
-
-        let (name, _) = res.expect("happy path returns Ok");
-        assert_eq!(name, "alpha", "first candidate should win");
-        assert_eq!(fake.calls(), 1, "only one backend call on happy path");
-    }
-
-    #[tokio::test]
-    async fn hedged_reducer_slow_first_hedges_to_second() {
-        let fake = FakeBackend::new(vec![
-            // alpha takes longer than hedge_delay; beta is fast.
-            (
-                "alpha",
-                FakeBehavior::OkAfter(Duration::from_millis(800), "alpha-late".into()),
-            ),
-            (
-                "beta",
-                FakeBehavior::OkAfter(Duration::from_millis(100), "beta-fast".into()),
-            ),
-        ]);
-        let backends: Vec<std::sync::Arc<dyn ModelBackend>> = vec![fake.clone(), fake.clone()];
-        let candidates = vec![("alpha".into(), 0), ("beta".into(), 1)];
-
-        let res = hedged_reducer_call(
-            &backends,
-            candidates,
-            vec![],
-            None,
-            Duration::from_secs(15),
-            Duration::from_millis(100),
-        )
-        .await;
-
-        let (name, body) = res.expect("hedge returns Ok");
-        assert_eq!(
-            name, "beta",
-            "hedge winner should be the faster second candidate"
-        );
-        assert_eq!(body, "beta-fast");
-        assert_eq!(fake.calls(), 2, "both candidates should have been issued");
-    }
-
-    #[tokio::test]
-    async fn hedged_reducer_fast_fail_starts_next_immediately() {
-        let fake = FakeBackend::new(vec![
-            (
-                "alpha",
-                FakeBehavior::ErrAfter(Duration::from_millis(50), "boom".into()),
-            ),
-            (
-                "beta",
-                FakeBehavior::OkAfter(Duration::from_millis(100), "beta-ok".into()),
-            ),
-        ]);
-        let backends: Vec<std::sync::Arc<dyn ModelBackend>> = vec![fake.clone(), fake.clone()];
-        let candidates = vec![("alpha".into(), 0), ("beta".into(), 1)];
-
-        let start = tokio::time::Instant::now();
-        let res = hedged_reducer_call(
-            &backends,
-            candidates,
-            vec![],
-            None,
-            Duration::from_secs(15),
-            // Large hedge_delay — the fast-fail path must not wait for it.
-            Duration::from_secs(60),
-        )
-        .await;
-
-        let (name, body) = res.expect("fail-then-recover returns Ok");
-        let elapsed = start.elapsed();
-        assert_eq!(name, "beta");
-        assert_eq!(body, "beta-ok");
-        assert_eq!(fake.calls(), 2);
-        assert!(
-            elapsed < Duration::from_secs(10),
-            "fast-fail should not wait for hedge_delay; took {elapsed:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn hedged_reducer_all_fail_returns_last_err() {
-        let fake = FakeBackend::new(vec![
-            (
-                "alpha",
-                FakeBehavior::ErrAfter(Duration::from_millis(10), "alpha-boom".into()),
-            ),
-            (
-                "beta",
-                FakeBehavior::ErrAfter(Duration::from_millis(10), "beta-boom".into()),
-            ),
-        ]);
-        let backends: Vec<std::sync::Arc<dyn ModelBackend>> = vec![fake.clone(), fake.clone()];
-        let candidates = vec![("alpha".into(), 0), ("beta".into(), 1)];
-
-        let res = hedged_reducer_call(
-            &backends,
-            candidates,
-            vec![],
-            None,
-            Duration::from_secs(15),
-            Duration::from_millis(200),
-        )
-        .await;
-
-        let err = res.expect_err("all-fail returns Err");
-        assert!(
-            err.contains("boom"),
-            "should surface a backend error: {err}"
-        );
-        assert_eq!(fake.calls(), 2);
-    }
 }
